+++ src/HTTP_Server_Basic.cpp
@@ -179,16 +179,16 @@ void HTTP_Server::start() {
   server.on("/hrslider", []() {
     String value = server.arg("value");
     if (value == "enable") {
-      rtConfig.setSimulateHr(true);
+      rtConfig.hr.setSimulate(true);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "HR Simulator turned on");
     } else if (value == "disable") {
-      rtConfig.setSimulateHr(false);
+      rtConfig.hr.setSimulate(false);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "HR Simulator turned off");
     } else {
-      rtConfig.setSimulatedHr(value.toInt());
-      SS2K_LOG(HTTP_SERVER_LOG_TAG, "HR is now: %d", rtConfig.getSimulatedHr());
+      rtConfig.hr.setValue(value.toInt());
+      SS2K_LOG(HTTP_SERVER_LOG_TAG, "HR is now: %d", rtConfig.hr.getValue());
       server.send(200, "text/plain", "OK");
     }
   });
@@ -196,16 +196,16 @@ void HTTP_Server::start() {
   server.on("/wattsslider", []() {
     String value = server.arg("value");
     if (value == "enable") {
-      rtConfig.setSimulateWatts(true);
+      rtConfig.watts.setSimulate(true);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "Watt Simulator turned on");
     } else if (value == "disable") {
-      rtConfig.setSimulateWatts(false);
+      rtConfig.watts.setSimulate(false);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "Watt Simulator turned off");
     } else {
-      rtConfig.setSimulatedWatts(value.toInt());
-      SS2K_LOG(HTTP_SERVER_LOG_TAG, "Watts are now: %d", rtConfig.getSimulatedWatts().value);
+      rtConfig.watts.setValue(value.toInt());
+      SS2K_LOG(HTTP_SERVER_LOG_TAG, "Watts are now: %d", rtConfig.watts.getValue());
       server.send(200, "text/plain", "OK");
     }
   });
@@ -213,16 +213,16 @@ void HTTP_Server::start() {
   server.on("/cadslider", []() {
     String value = server.arg("value");
     if (value == "enable") {
-      rtConfig.setSimulateCad(true);
+      rtConfig.cad.setSimulate(true);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "CAD Simulator turned on");
     } else if (value == "disable") {
-      rtConfig.setSimulateCad(false);
+      rtConfig.cad.setSimulate(false);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "CAD Simulator turned off");
     } else {
-      rtConfig.setSimulatedCad(value.toInt());
-      SS2K_LOG(HTTP_SERVER_LOG_TAG, "CAD is now: %d", rtConfig.getSimulatedCad());
+      rtConfig.cad.setValue(value.toInt());
+      SS2K_LOG(HTTP_SERVER_LOG_TAG, "CAD is now: %d", rtConfig.cad.getValue());
       server.send(200, "text/plain", "OK");
     }
   });
@@ -243,16 +243,16 @@ void HTTP_Server::start() {
   server.on("/targetwattsslider", []() {
     String value = server.arg("value");
     if (value == "enable") {
-      rtConfig.setSimulateTargetWatts(true);
+      rtConfig.setSimTargetWatts(true);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "Target Watts Simulator turned on");
     } else if (value == "disable") {
-      rtConfig.setSimulateTargetWatts(false);
+      rtConfig.setSimTargetWatts(false);
       server.send(200, "text/plain", "OK");
       SS2K_LOG(HTTP_SERVER_LOG_TAG, "Target Watts Simulator turned off");
     } else {
-      rtConfig.setTargetWatts(value.toInt());
-      SS2K_LOG(HTTP_SERVER_LOG_TAG, "Target Watts are now: %d", rtConfig.getTargetWatts());
+      rtConfig.watts.setTarget(value.toInt());
+      SS2K_LOG(HTTP_SERVER_LOG_TAG, "Target Watts are now: %d", rtConfig.watts.getTarget());
       server.send(200, "text/plain", "OK");
     }
   });