+++ CHANGELOG.md
@@ -119,6 +119,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added WebSockets for logging [see #173](https://github.com/doudar/SmartSpin2k/issues/173)
 - Reworked logging to run log-appender outside the worker task (task no longer blocked by logger traffic).
 - WebsocketsAppender can handle multiple (up to 4) clients. Status.html will reconnect if connection to websockets server is disconnected.
+- Added ability to set travel limits based on resistance feedback from a bike.
 
 ### Changed
 - Refactored ERG.
@@ -142,6 +143,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increased JSON size for userConfig (hopefully fix config saving issues). 
 - Changed LOGE messages in spiffs logging to regular LOG messages so they will display via network logging. 
 - Complete BLE Client connection code rebase. 
+- Refactored rtConfig to use more measurement class. 
+- Increased stepper speed when a Peloton is connected. (very light resistance)
 
 ### Fixed
 - bluetoothscanner.html now lists fitness machine services in the PM list.